add borders

+++ nvim/init.lua
@@ -5,7 +5,7 @@ local opt = vim.opt
 opt.syntax = "enable"
 opt.number = true
 opt.relativenumber = true
-vim.o.winborder = "single"
+vim.o.winborder = "double"
 
 opt.colorcolumn = "0"
 opt.wrap = false
